Extracted World envelope from GraphMetadata
